+++ packages/components-js/tests/vrt/specs/states/marque.vrt.ts
@@ -6,7 +6,7 @@ import {
   type PrefersColorScheme,
   setContentWithDesignSystem,
 } from '../../helpers';
-import { type Theme } from '@porsche-design-system/utilities-v2';
+import { type Theme } from '@porsche-design-system/styles';
 import { viewportWidthM } from '@porsche-design-system/shared/testing/playwright.vrt';
 
 const component = 'marque';