+++ packages/components-js/tests/vrt/specs/states/table.vrt.ts
@@ -9,7 +9,7 @@ import {
   setContentWithDesignSystem,
   setSortToAllTableHeadCell,
 } from '../../helpers';
-import { type Theme } from '@porsche-design-system/utilities-v2';
+import { type Theme } from '@porsche-design-system/styles';
 
 const component = 'table';