+++ packages/components/src/components/canvas/canvas-styles.ts
@@ -1,6 +1,6 @@
 import { getCss } from '../../utils';
 import { addImportantToEachRule, colorSchemeStyles, getTransition, hostHiddenStyles } from '../../styles';
-import { getMediaQueryMin, spacingStaticXSmall, gridGap } from '@porsche-design-system/utilities-v2';
+import { getMediaQueryMin, spacingStaticXSmall, gridGap } from '@porsche-design-system/styles';
 import { type CanvasSidebarWidth, type CanvasSidebarStartWidth, type CanvasSidebarEndWidth } from './canvas-utils';
 
 const cssVariableSidebarStartWidth = '--p-canvas-sidebar-start-width';