+++ packages/components/src/components/select-wrapper/select-wrapper-dropdown/select-wrapper-dropdown-styles.ts
@@ -18,7 +18,7 @@ import {
   spacingStaticSmall,
   textSmallStyle,
   fontSizeTextXSmall
-} from '@porsche-design-system/utilities-v2';
+} from '@porsche-design-system/styles';
 
 import { getNoResultsOptionJssStyle, OPTION_HEIGHT } from '../../../styles/option-styles';
 import { getThemedFormStateColors } from '../../../styles/form-state-color-styles';