+++ packages/utilities/projects/utilities/src/js/theme/themeLightBackgroundFrosted.ts
@@ -1,0 +1,1 @@
+export const themeLightBackgroundFrosted = 'hsl(240 4% 85%/35%)';