+++ packages/components/CHANGELOG.md
@@ -14,6 +14,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### [Unreleased]
 
+#### Added
+
+- Styles: `theme{Light|Dark}BackgroundFrosted` and `$pds-theme-{light|dark}-background-frosted` color
+  ([#3409](https://github.com/porsche-design-system/porsche-design-system/pull/3409))
+
 #### Changed
 
 - `Icon`: All icons are up-to-date with the One UI look